jmap_calendar: refactor jmap context to jmap_ical module
